feat(buttons): add .btn-login
The .btn-login class is used to improve the semantics of red buttons in
the design language of Canon. Red buttons are used for logging in in
Canon, as opposed to dangerous actions in Bootstrap. Canon-bootstrap
will have both as we are exploring new use cases for buttons at the same
time as supporting the Canon language.

This change also includes a component demo. The intent is if a new use
case for the framework doesn't fit well in the existing realistic demo,
you can put it by itself in the component demo.